refactor: use non-primitive types for Product quantity, price, discount
Change `int` → `Integer`, `double price/discount` → `BigDecimal` in
Product entity and ProductDTO. Update tests to use BigDecimal.valueOf().

AI-Tool: claude (claude-sonnet-4-6)

+++ src/main/java/com/ecommerce/project/dto/ProductDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,8 +13,8 @@ public class ProductDTO {
     private Long id;
     private String name;
     private String description;
-    private int quantity;
-    private double price;
-    private double discount;
+    private Integer quantity;
+    private BigDecimal price;
+    private BigDecimal discount;
     private CategoryDTO category;
 }
